Small refactor
As I am learning C++ syntax I am slowly beginning to refactor my code to make it more sensible. I am slowly working towards being able to instantiate an instance of the BlitRPG::Map class with a standardised set of params (maybe a JSON file). It is my intention then to use that Map Class as an interface to all Map based functions.

It will also be passed to the rendering layer as it has all of the necessary information required to render the screen.